requirements.txt
FundamentalAnalysis
pandas

django
djangorestframework